fix: isLoading 추가를 통한 data 반환 수정
