--- src/pages/statistic/DummyData.js
@@ -1,69 +1,0 @@
-/**
- * 임시 목업 데이터
- * */
-export const yearData = {
-  year: 2022,
-  income: [
-    {
-      name: '월급',
-      percent: 50,
-      total: 500000,
-    },
-    {
-      name: '주식',
-      percent: 30,
-      total: 300000,
-    },
-  ],
-  expenditure: [
-    {
-      name: '식비',
-      percent: 50,
-      total: 50000,
-    },
-    {
-      name: '패션/미용',
-      percent: 30,
-      total: 30000,
-    },
-    {
-      name: '교육',
-      percent: 20,
-      total: 20000,
-    },
-  ],
-};
-
-export const monthData = {
-  year: 2022,
-  month: 7,
-  income: [
-    {
-      name: '월급',
-      percent: 50,
-      total: 500000,
-    },
-    {
-      name: '주식',
-      percent: 30,
-      total: 300000,
-    },
-  ],
-  expenditure: [
-    {
-      name: '식비',
-      percent: 50,
-      total: 50000,
-    },
-    {
-      name: '패션/미용',
-      percent: 30,
-      total: 30000,
-    },
-    {
-      name: '교육',
-      percent: 20,
-      total: 20000,
-    },
-  ],
-};
