+++ src/types/accoutBook.d.ts
@@ -41,18 +41,3 @@ declare type MonthlyAccount = {
   totalSum: number;
   month: number;
 };
-
-declare type StatisticIncome = {
-  name: string;
-  total: number;
-  percent: number;
-};
-
-declare type StatisticResults = {
-  year: number;
-  month: number | null;
-  incomeTotalSum: number;
-  expenditureTotalSum: number;
-  incomes: StatisticIncome[];
-  expenditures: StatisticIncome[];
-};
